+++ docs/changelog.rst
@@ -25,6 +25,7 @@ Changelog
 * Improved Project Browser to not show irrelevant items
 * Improved Test Level/Start Game button behavior when game failed to launch
 * Level Editor now saves aggregate logs to disk. User can browse logs folder from Help > Browse Logs...
+* New Project dialog no longer allows selecting non-empty folders for new projects
 * Unified Engine and Run menubar items into a single Debug menubar item
 
 0.38.0